added speed calculation for clone creation and deletion (#3092)
* added speed calculation for clone creation and deletion
